+++ upgrade/constants.go
@@ -1,0 +1,12 @@
+package upgrade
+
+// UpgradeName is the name of this specific software upgrade used on-chain.
+const UpgradeName = "helium"
+
+const UpgradeASCII = `
+ _   _      _ _                 
+| | | | ___| (_)_   _ _ __ ___  
+| |_| |/ _ \ | | | | |  _   _ \ 
+|  _  |  __/ | | |_| | | | | | |
+|_| |_|\___|_|_|\__,_|_| |_| |_|
+`