+++ x/globalfee/alias.go
@@ -1,9 +1,0 @@
-package globalfee
-
-import (
-	"github.com/noble-assets/noble/v7/x/globalfee/types"
-)
-
-const (
-	ModuleName = types.ModuleName
-)